reword word_count instructions
The current instructions say that if the *input* does contain ignored characters, the function should return an empty dictionary. However, implementing this fails the test on line 12; "Hello     hello" contains no ignored characters, so according to the current instructions, an empty dictionary should be returned. My proposed amendment is what I believe the instructions are intending to say.